Decoupled crane logic from crane command line code   - Added commands package   - Added api package   - Extracted logic from commands into api package
