+++ pkg/crane/api/export.go
@@ -12,49 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package crane
+package api
 
 import (
-	"io"
-	"log"
-	"os"
-
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
 	"github.com/google/go-containerregistry/pkg/v1/remote"
-	"github.com/spf13/cobra"
+	"io"
+	"log"
+	"os"
 )
 
-func init() { Root.AddCommand(NewCmdExport()) }
-
-// NewCmdExport creates a new cobra.Command for the export subcommand.
-func NewCmdExport() *cobra.Command {
-	exportCmd := &cobra.Command{
-		Use:   "export IMAGE OUTPUT",
-		Short: "Export contents of a remote image as a tarball",
-		Example: `  # Write tarball to stdout
-  crane export ubuntu -
-
-  # Write tarball to file
-  crane export ubuntu ubuntu.tar`,
-		Args: cobra.ExactArgs(2),
-		Run: func(_ *cobra.Command, args []string) {
-			doExport(args[0], args[1])
-		},
-	}
-
-	return exportCmd
-}
-
-func openFile(s string) (*os.File, error) {
-	if s == "-" {
-		return os.Stdout, nil
-	}
-	return os.Create(s)
-}
-
-func doExport(src, dst string) {
+func Export(src string, dst string) {
 	srcRef, err := name.ParseReference(src)
 	if err != nil {
 		log.Fatalf("parsing reference %q: %v", src, err)
@@ -75,4 +45,11 @@ func doExport(src, dst string) {
 	if _, err := io.Copy(out, fs); err != nil {
 		log.Fatal(err)
 	}
+}
+
+func openFile(s string) (*os.File, error) {
+	if s == "-" {
+		return os.Stdout, nil
+	}
+	return os.Create(s)
 }
