+++ pkg/crane/api/get.go
@@ -12,44 +12,33 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package crane
+package api
 
 import (
-	"log"
+	"fmt"
 
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
+	v1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/remote"
-	"github.com/google/go-containerregistry/pkg/v1/tarball"
-	"github.com/spf13/cobra"
 )
 
-func init() { Root.AddCommand(NewCmdPush()) }
-
-// NewCmdPush creates a new cobra.Command for the push subcommand.
-func NewCmdPush() *cobra.Command {
-	return &cobra.Command{
-		Use:   "push",
-		Short: "Push image contents as a tarball to a remote registry",
-		Args:  cobra.ExactArgs(2),
-		Run:   push,
+func getImage(refStr string) (v1.Image, name.Reference, error) {
+	reference, err := name.ParseReference(refStr)
+	if err != nil {
+		return nil, nil, fmt.Errorf("parsing reference %q: %v", refStr, err)
 	}
+	img, err := remote.Image(reference, remote.WithAuthFromKeychain(authn.DefaultKeychain))
+	if err != nil {
+		return nil, nil, fmt.Errorf("reading image %q: %v", reference, err)
+	}
+	return img, reference, nil
 }
 
-func push(_ *cobra.Command, args []string) {
-	src, dst := args[0], args[1]
-	t, err := name.NewTag(dst)
+func getManifest(refStr string) (*remote.Descriptor, error) {
+	ref, err := name.ParseReference(refStr)
 	if err != nil {
-		log.Fatalf("parsing tag %q: %v", dst, err)
+		return nil, fmt.Errorf("parsing reference %q: %v", refStr, err)
 	}
-	log.Printf("Pushing %v", t)
-
-	i, err := tarball.ImageFromPath(src, nil)
-	if err != nil {
-		log.Fatalf("reading image %q: %v", src, err)
-	}
-
-	if err := remote.Write(t, i, remote.WithAuthFromKeychain(authn.DefaultKeychain)); err != nil {
-		log.Fatalf("writing image %q: %v", t, err)
-	}
+	return remote.Get(ref, remote.WithAuthFromKeychain(authn.DefaultKeychain))
 }
