+++ pkg/crane/api/manifest.go
@@ -1,3 +1,5 @@
+// Copyright 2018 Google LLC All Rights Reserved.
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -10,14 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package crane
+package api
 
-import "github.com/spf13/cobra"
+import (
+	"fmt"
+	"log"
+)
 
-// Root is the top-level cobra.Command for crane.
-var Root = &cobra.Command{
-	Use:               "crane",
-	Short:             "Crane is a tool for managing container images",
-	Run:               func(cmd *cobra.Command, _ []string) { cmd.Usage() },
-	DisableAutoGenTag: true,
+func Manifest(ref string) {
+	desc, err := getManifest(ref)
+	if err != nil {
+		log.Fatal(err)
+	}
+	fmt.Print(string(desc.Manifest))
 }
