+++ pkg/crane/api/validate.go
@@ -12,12 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package crane
+package api
 
 import (
 	"fmt"
-	"log"
-
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
@@ -25,30 +23,10 @@ import (
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/google/go-containerregistry/pkg/v1/tarball"
 	"github.com/google/go-containerregistry/pkg/v1/validate"
-	"github.com/spf13/cobra"
+	"log"
 )
 
-func init() { Root.AddCommand(NewCmdValidate()) }
-
-// NewCmdValidate creates a new cobra.Command for the validate subcommand.
-func NewCmdValidate() *cobra.Command {
-	var tarballPath, remoteRef, daemonRef string
-	validateCmd := &cobra.Command{
-		Use:   "validate",
-		Short: "Validate that an image is well-formed",
-		Args:  cobra.ExactArgs(0),
-		Run: func(_ *cobra.Command, args []string) {
-			doValidate(tarballPath, remoteRef, daemonRef)
-		},
-	}
-	validateCmd.Flags().StringVar(&tarballPath, "tarball", "", "Path to tarball to validate")
-	validateCmd.Flags().StringVar(&remoteRef, "remote", "", "Name of remote image to validate")
-	validateCmd.Flags().StringVar(&daemonRef, "daemon", "", "Name of image in daemon to validate")
-
-	return validateCmd
-}
-
-func doValidate(tarballPath, remoteRef, daemonRef string) {
+func Validate(tarballPath string, remoteRef string, daemonRef string) {
 	for flag, maker := range map[string]func(string) (v1.Image, error){
 		tarballPath: makeTarball,
 		remoteRef:   makeRemote,
